+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 =======
 
+0.6.0 (2015-06-??)
+---------------------
+
+* Support for Django 1.6 and lower is now deprecated.
+* Improved test harness now tests coverage and pep8
+
 0.5.0 (2015-05-25)
 ---------------------